schemify improvements and repairs
These changes seem to improve generated code for Chez, but they
also seem to make Chez's compiler take 20% longer, so that the
time to expand `racket/base` goes up (with a corresponding jump in
allocation). Maybe something in the generated code is not as good,
or maybe slightly better code gives the compiler more to work with.