new HAMT implementation:
- Specialized versions for immutable hashes and sets. Sets
do not store values. Hashes store keys and values directly
in the node array, instead of in an `entry` struct.

- Procedures do not pass equality predicates and hash-code
functions. Instead, they pass a symbolic `eqtype`, and we
dispatch on that to use the correct predicate/hash-code.
(We could also specialize all of the procedures for their
eqtype. I did this is a previous branch. It is a bit faster
but perhaps not enough to justify so much code duplication.

- Significantly faster than existing version on large table
examples in hash-demo.ss, but slower for small table examples.
My *guess* is that the existing specialization for #t values
accounts for that, since all of the values in those examples
use #t values. However, I haven't verified that this is the
reason for the performance difference. The new specialized
set-hamt is faster than either hash implementation on those
examples.